[#13] refactor: 롤링 배너 util 함수 분리 & 롤링 방향 교차 구현
