+++ src/app/globals.css
@@ -14,6 +14,12 @@
       transform: translateX(-50%);
     }
   }
+  
+  --animate-rolling-reverse: rolling-reverse 20s linear infinite;
+  @keyframes rolling-reverse {
+    0% { transform: translateX(-50%); }
+    100% { transform: translateX(0); }
+  }
 }
 
 body {
